refactor/updated the code according to the comments
